文档增加ClickGap和TimeLog注解的说明

+++ annotation/src/main/java/com/catchpig/annotation/ClickGap.kt
@@ -13,5 +13,5 @@ annotation class ClickGap(
         /**
          * 重复点击间隔时间
          */
-        val value: Int = 800
+        val value: Long = 800
 )
